Rename: ApplyStatusItem -> StatusItem 이름 변경
