+++ app/(route)/mypage/(components)/StatusItem.tsx
